product Review

+++ database/migrations/2021_08_04_174041_create_areas_table.php
@@ -15,7 +15,7 @@ class CreateAreasTable extends Migration
     {
         Schema::create('areas', function (Blueprint $table) {
             $table->id();
-            $table->integer('district_id');
+            $table->unsignedBigInteger('district_id');
             $table->string('area');
             $table->timestamps();
         });
